AMMI-based indexes for stability and simultaneous selection

This function computes the following AMMI-based stability indexes: ASV, AMMI stability value (Purchase et al., 2000); SIPC, sums of the absolute value of the IPCA scores (Sneller et al. 1997); EV, averages of the squared eigenvector values (Sneller et al. 1997); and Za, absolute value of the relative contribution of IPCAs to the interaction (Zali et al. 2012).

AMMI_indexes(x, order.y = NULL)

Arguments

x
An object of class WAAS.AMMI

order.y
A vector of the same length of x used to order the response variable. Each element of the vector must be one of the "h" or "l". If "h" is used, the response variable will be ordered from maximum to minimum. If "l" is used then the response variable will be ordered from minimum to maximum.

Details

The ASV index is computed as follows: $$AS{V_i} = {\left[ {{{\left[ {\frac{{r\mathop \lambda \nolimits_1^2 }}{{r\mathop \lambda \nolimits_2^2 }} \times (\mathop \lambda \nolimits_1^{0.5} {a_{i1}}{t_{j1}})} \right]}^2} + {{(\mathop \lambda \nolimits_2^{0.5} {a_{i2}}{t_{j2}})}^2}} \right]^{0.5}}$$

where \(r\) is the number of replications included in the analysis,

The SIPC index is computed as follows: $$SIP{C_i} = \sum\nolimits_{k = 1}^P {\left| {\mathop {|\lambda }\nolimits_k^{0.5} {a_{ik}}} \right|}$$

where \(P\) is the number of IPCA retained via F-tests.

The EV index is computed as follows: $$E{V_i} = \sum\nolimits_{k = 1}^P {\mathop a\nolimits_{ik}^2 } /P$$

The ZA index is computed as follows: $$Z{a_i} = \sum\nolimits_{k = 1}^P {{\theta _k}{a_{ik}}} $$

where \(\theta _k\) is the percentage sum of squares explained by the kth IPCA.

Four simultaneous selection indexes (ssi) are also computed by summation of the ranks of the ASV, SIPC, EV and Za indexes and the ranks of the mean yields (Farshadfar, 2008), which results in ssiASV, ssiSIPC, ssiEV, and ssiZa, respectively.

Value

A dataframe contaning the indexes

References

Purchase, J.L., H. Hatting, and C.S. van Deventer. 2000. Genotype vs environment interaction of winter wheat (Triticum aestivum L.) in South Africa: II. Stability analysis of yield performance. South African J. Plant Soil 17:101-107. doi:10.1080/02571862.2000.10634878

Sneller, C.H., L. Kilgore-Norquest, and D. Dombek. 1997. Repeatability of Yield Stability Statistics in Soybean. Crop Sci. 37:383-390. doi:10.2135/cropsci1997.0011183X003700020013x

Zali, H., E. Farshadfar, S.H. Sabaghpour, and R. Karimizadeh. 2012. Evaluation of genotype vs environment interaction in chickpea using measures of stability from AMMI model. Ann. Biol. Res. 3:3126-3136. http://eprints.icrisat.ac.in/id/eprint/7173

Examples

library(METAAB)
model = WAAS.AMMI(data_ge,
                  resp = c(GY, HM),
                  gen = GEN,
                  env = ENV,
                  rep = REP)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

model_indexes = AMMI_indexes(model)