Nonparametric confidence interval for Pearson's correlation coefficient

Computes the half-width confidence interval for correlation coefficient based on the nonparametri method proposed by Olivoto et al. (2018).

corr.ci(data = NA, r = NULL, n = NULL)

Arguments

data
A dataset containing variables only or a symetric correlation matrix.

r
If data is not available, provide the value for correlation coefficient.

n
The sample size if data is a correlation matrix or if r is informed.

Details

The half-width confidence interval is computed according to the following equation: $$CI_w = 0.45304^r \times 2.25152 \times n^{-0.50089}$$

where \(n\) is the sample size and r is the correlation coefficient.

References

Olivoto, T., A.D.C. Lucio, V.Q. Souza, M. Nardino, M.I. Diel, B.G. Sari, D.. K. Krysczun, D. Meira, and C. Meier. 2018. Confidence interval width for Pearson's correlation coefficient: a Gaussian-independent estimator based on sample size and strength of association. Agron. J. 110:1-8. doi:10.2134/agronj2017.09.0566.

Examples

corr.ci(iris[,1:3])

#>                                   Corr         CI         LL          UL
#> Sepal.Length x Sepal.Width  -0.1175698 0.16674987 -0.2843197  0.04918009
#> Sepal.Length x Petal.Length  0.8717538 0.09177601  0.7799778  0.96352979
#> Sepal.Width x Petal.Length  -0.4284401 0.13036727 -0.5588074 -0.29807284