A replicate-based data of 10 genotypes cultivated in 14 environments

This dataset contain data on two variables assessed in 10 genotypes growing in in 11 environments. The experimental design was a RCBD with 3 replicates(blocks). This data provide examples for several functions of METAAB package.

Arguments

ENV
A factor with 14 levels; each level represents one cultivation environment.

GEN
A factor with 10 levels; each level represents one genotype.

REP
A factor with 3 levels; each level represents one replication/block.

GY
A continuous variable (grain yield) observed in each plot.

HM
A continuous variable (hectoliter mass) observed in each plot.