Make a two-way table based on categorical and numerical arguments

This function help users to easely make a two-way table from a "long format" data.

mmat(data, row, col, value)

Arguments

data
The dataset. Must contains at least two categorical columns.

row
The column of data in which the mean of each level will correspond to one line in the output.

col
The column of data in which the mean of each level will correspond to one column in the output.

value
The column of data that contains the values to fill the two-way table.

Examples

datage = expand.grid(ENV = paste("E", 1:10, sep = ""),
                     GEN = paste("G", 1:20, sep = ""),
                     REP = paste(c("I", "II", "III")))
datage$Y = rnorm(600, 50, 20)
matrix = mmat(datage, row = GEN, col = ENV, val = Y)
matrix

#>           E1       E2       E3       E4       E5       E6       E7       E8
#> G1  49.37792 55.13687 26.86872 54.93177 27.85879 50.19386 34.73272 58.13833
#> G2  38.54974 47.54559 64.54569 27.46606 65.82683 42.32438 53.67220 40.74286
#> G3  54.73357 50.73037 46.51085 49.09756 36.34063 62.36640 41.37715 34.11481
#> G4  60.84411 36.75305 60.37944 58.08828 46.17929 41.58140 23.85776 49.44226
#> G5  35.06387 59.25126 48.72921 53.89305 46.50404 75.43664 41.74962 46.46521
#> G6  62.07543 47.59868 34.54384 34.55412 59.34229 34.24102 36.79485 62.31155
#> G7  64.94803 34.25248 68.41461 41.78521 29.94772 27.64309 65.76679 43.60839
#> G8  36.40383 31.99226 52.39314 37.86502 49.66895 62.72347 64.55902 52.88366
#> G9  67.82920 56.87818 63.74018 61.83512 58.71812 60.53978 29.88967 55.39298
#> G10 52.09763 53.84140 47.27893 43.54614 60.56092 49.25397 30.96307 64.56856
#> G11 40.30092 42.08188 52.07673 53.89328 31.59034 42.46160 54.96397 64.28507
#> G12 36.74856 50.57351 54.40484 32.10242 55.09038 58.49455 66.72528 37.58964
#> G13 63.06083 54.02148 72.65474 56.99202 68.55809 35.91037 57.12614 38.82211
#> G14 44.94567 46.61960 55.21905 41.97309 48.78396 66.29785 40.82194 36.42788
#> G15 59.44293 57.83702 70.88862 41.64106 47.47358 65.61784 39.03675 54.55872
#> G16 43.77709 43.49035 57.87468 60.16360 68.40200 59.26617 48.40105 57.58096
#> G17 38.95419 62.32172 41.95457 73.31722 19.72751 44.41158 26.30806 42.20715
#> G18 44.90854 29.81715 46.61770 45.67361 58.30730 66.60716 61.58867 32.39811
#> G19 41.76389 39.40835 21.53838 61.51729 58.60537 72.98428 54.12915 81.08515
#> G20 64.87630 51.39179 56.42256 37.68514 56.85594 63.72094 39.82656 66.57126
#>           E9      E10
#> G1  30.52273 54.87726
#> G2  49.89529 51.66780
#> G3  49.72176 35.55967
#> G4  56.36852 34.08503
#> G5  63.82475 40.42737
#> G6  51.20206 59.68185
#> G7  60.43873 44.54364
#> G8  59.50105 38.45377
#> G9  74.97079 31.16163
#> G10 60.70191 71.19628
#> G11 70.68821 53.56056
#> G12 36.22826 65.31534
#> G13 42.79836 50.42880
#> G14 58.17120 63.65424
#> G15 43.30579 60.82516
#> G16 62.48800 45.29835
#> G17 59.01660 50.18345
#> G18 46.76467 69.78182
#> G19 51.60204 56.43463
#> G20 61.91270 38.78349