Partial correlation coefficient

Estimates of partial correlation coefficient based on data raw or correlation matrices.

partial.corr(data, n = NULL, method = "pearson")

Arguments

data
The data, must be a dataframe or a symetric correlation matrix.

n
The number of subjects used if data is a correlation matrix.

method
If data is a correlation matrix, then codemethod is the method for computing the correlation; one of the "pearson" (default), "kendall", or "spearman"

Value

linear.mat

The matrix of linear correlation.

partial.mat

The matrix of partial correlations.

results

Hypotesis testing for each pairwise comparation

call

The function call

Examples

partial.corr(iris[,1:4])

#> $linear.mat
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    1.0000000  -0.1175698    0.8717538   0.8179411
#> Sepal.Width    -0.1175698   1.0000000   -0.4284401  -0.3661259
#> Petal.Length    0.8717538  -0.4284401    1.0000000   0.9628654
#> Petal.Width     0.8179411  -0.3661259    0.9628654   1.0000000
#> 
#> $partial.mat
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    1.0000000   0.6285707    0.7190656  -0.3396174
#> Sepal.Width     0.6285707   1.0000000   -0.6152919   0.3526260
#> Petal.Length    0.7190656  -0.6152919    1.0000000   0.8707698
#> Petal.Width    -0.3396174   0.3526260    0.8707698   1.0000000
#> 
#> $results
#>                                 linear    partial         t         prob
#> Sepal.Length x Sepal.Width  -0.1175698  0.6285707  9.765380 0.000000e+00
#> Sepal.Length x Petal.Length  0.8717538  0.7190656 12.502483 0.000000e+00
#> Sepal.Length x Petal.Width   0.8179411 -0.3396174 -4.362929 2.412876e-05
#> Sepal.Width x Petal.Length  -0.4284401 -0.6152919 -9.431189 0.000000e+00
#> Sepal.Width x Petal.Width   -0.3661259  0.3526260  4.553279 1.104958e-05
#> Petal.Length x Petal.Width   0.9628654  0.8707698 21.398708 0.000000e+00
#> 
#> $call
#> partial.corr(data = iris[, 1:4])
#>