Path coefficients with minimal multicollinearity

Estimates of direct and indirect effects. An algorithm to select a set of predictors with minimal multicollinearity and high explanatory power is implemented.

path_coeff(.data, resp, pred = NULL, exclude = FALSE,
           correction = NULL, knumber = 50, brutstep = FALSE,
           maxvif = 10, missingval = "pairwise.complete.obs",
           verbose = TRUE)

Arguments

.data
The data. Must be a dataframe or an object of class group_factors.

resp
The dependent variable.

pred
The predictor variables, set to NULL, i.e., the predictor variables are all the numeric variables in the data except that in resp.

exclude
Logical argument, set to false. If exclude = TRUE, then the variables in pred are deleted from the data, and the analysis will use as predictor those that remained, except that in resp.

correction
Set to NULL. A correction value (k) that will be added into the diagonal elements of the X'X matrix aiming at reducing the multicollinearity problems (Olivoto et al., 2017). Also as known as ridge regression.

knumber
When correction = NULL, a plot showing the values of direct effects in a set of different k values (0-1) is produced. knumber is the number of k values used in the range of 0 to 1.

brutstep
Logical argument, set to FALSE. If true, then an algorithm will select a subset of variables with minimal multicollinearity and fit a set of possible models. See the Details section for more information.

maxvif
The maximum value for the Variance Inflaction Factor (cut point) that will be accepted. See the Details section for more information.

missingval
How to deal with missing values. For more information, please see ?cor.

verbose
If verbose = TRUE then some results are shown in the console.

Value

Corr.x

A correlation matrix between the predictor variables.

Corr.y

A vector of correlations between each predictor variable with the dependent variable.

Coefficients

The path coefficients. Direct effects are the diagonal elements, and the indirect effects those in the off-diagonal elements (column)

Eigen

Eigenvectors and eigenvalues of the Corr.x.

VIF

The Variance Inflaction Factors.

plot

A ggplot2-based graphic showing the direct effects in 21 different k values..

Predictors

The predictor variables used in the model.

CN

The Condition Number, i.e., the ratio between the highest and lowest eigenvalue.

Det

The matrix determinant of the Corr.x..

R2

The coefficient of determination of the model.

Residual

The residual effect of the model.

Response

The response variable.

weightvar

The order of the predictor variables with the higest weigth (highest eigenvector) in the lowest eigenvalue.

Details

When brutstep = TRUE, first, the algorithm will select a set of predictors with minimal multicollinearity. The selection is based on the variance inflation factor (VIF). An iterative process is performed until the maximum VIF observed is less than maxvif. The variables selected in this iterative process are then used in a series of stepwise-based regressions. The first model is fitted and p-1 predictor variables are retained (p is the number of variables selected in the iterative process. The second model adjusts a regression considering p-2 selected variables, and so on until the last model, which considers only two variables. Three objects are created. Summary, with the process summary, Models, containing the aforementioned values for all the adjusted models; and Selectedpred, a vector with the name of the selected variables in the iterative process.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. 10.2134/agronj2016.04.0196.

Examples

# Using the correlation matrix
library(METAAB)
library(dplyr)

pcoeff = data_ge2 %>%
         path_coeff(resp = KW, pred = c("TKW", "NKE"))

#> The factors ENV GEN REP PLA where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'group_factors() before.' 

#> Weak multicollinearity. 
#> Condition Number = 1.181
#> You will probably have path coefficients close to being unbiased. 

pcoeff2 = data_ge2 %>%
         path_coeff(resp = KW,
                    pred = c("PH", "EH", "NKE", "TKW"),
                    exclude = TRUE)

#> The factors ENV GEN REP PLA where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'group_factors() before.' 

#> Severe multicollinearity. 
#> Condition Number = 1757.354
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 

pcoeff3 = data_ge2 %>%
          path_coeff(resp = KW,
                     brutstep = TRUE,
                     maxvif = 5)

#> The factors ENV GEN REP PLA where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'group_factors() before.' 

#> -------------------------------------------------------------------------- 
#> The algorithm has selected a set of 9predictors with largest VIF = 3.374. 
#> Selected predictors: EP PERK CDED PH NKR EL NR TKW ED 
#> A forward stepwise-based selection procedure will fit 7 models. 
#> -------------------------------------------------------------------------- 
#> Adjusting the model 1 with 8 predictors (14.29% concluded)
#> Adjusting the model 2 with 7 predictors (28.57% concluded)
#> Adjusting the model 3 with 6 predictors (42.86% concluded)
#> Adjusting the model 4 with 5 predictors (57.14% concluded)
#> Adjusting the model 5 with 4 predictors (71.43% concluded)
#> Adjusting the model 6 with 3 predictors (85.71% concluded)
#> Adjusting the model 7 with 2 predictors (100% concluded)
#> Done! 
#> -------------------------------------------------------------------------- 
#> Summary of the adjusted models 
#> -------------------------------------------------------------------------- 
#>     Model      AIC Numpred        CN Determinant        R2  Residual   maxVIF
#> 2 Model 7 7077.035       2  1.836837  0.91298120 0.7769493 0.2230507 1.095313
#> 3 Model 6 6779.019       3  3.542543  0.68262402 0.8481697 0.1518303 1.449957
#> 4 Model 5 6686.519       4 10.287630  0.31386137 0.8654939 0.1345061 2.994049
#> 5 Model 4 6613.968       5 13.213486  0.15897502 0.8777545 0.1222455 3.016821
#> 6 Model 3 6564.740       6 13.609328  0.14204137 0.8855252 0.1144748 3.081659
#> 7 Model 2 6551.198       7 16.079825  0.09829292 0.8877836 0.1122164 3.101736
#> 8 Model 1 6551.346       8 17.062130  0.08371867 0.8880497 0.1119503 3.157106
#> --------------------------------------------------------------------------

pcoeff4 = data_ge2 %>%
          group_factors(ENV) %>%
          path_coeff(resp = KW)

#> The factors GEN REP PLA where ignored. Use 'keep_factors = TRUE' to keep this columns in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level A1 
#> ----------------------------------------------------------------------------
#> Severe multicollinearity. 
#> Condition Number = 3350.754
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 
#> 
#> ----------------------------------------------------------------------------
#> Level A2 
#> ----------------------------------------------------------------------------
#> Severe multicollinearity. 
#> Condition Number = 1511.159
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 
#> 
#> ----------------------------------------------------------------------------
#> Level A3 
#> ----------------------------------------------------------------------------
#> Severe multicollinearity. 
#> Condition Number = 3948.88
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 
#> 
#> ----------------------------------------------------------------------------
#> Level A4 
#> ----------------------------------------------------------------------------
#> Severe multicollinearity. 
#> Condition Number = 4304.863
#> Please, consider using a correction factor, or use 'brutstep = TRUE'.