Plot the RMSPD of all AMMI-family tested models

Boxplot showing the Root Means Square Prediction Difference of all AMMI family models obtained in a cross-validation procedure.

# S3 method for validation.AMMIF
plot(x, violin = FALSE, export = FALSE,  x.lab = NULL,
     y.lab = NULL, size.tex.lab = 12, file.type = "pdf",
     file.name = NULL, theme = theme_waasb(),width = 6,
     height = 6, resolution = 300, col.violin = "gray90",
     col.boxplot = "gray70", width.boxplot = 0.2,
     x.lim = NULL, x.breaks = waiver(), ...)

Arguments

x
An object of class validation.AMMIF.

violin
Define if a violine plot is used with boxplot. Default is "TRUE"

export
Export (or not) the plot. Default is T.

x.lab
The label of x-axis. New arguments can be inserted as x.lab = "my x label".

y.lab
The label of y-axis. New arguments can be inserted as y.lab = "my y label".

size.tex.lab
The size of the text in axis text and labels.

file.type
The type of file to be exported. Default is pdf, Graphic can also be exported in *.tiff format by declaring file.type = "tiff".

file.name
The name of the file for exportation, default is NULL, i.e. the files are automatically named.

theme
The graphical theme of the plot. Default is `theme = theme_waasb()`. Please, see `?WAASB::theme_waasb`. An own theme can be applied using the arguments: `theme = theme_waasb() + theme(some stuff here)`. For more details, please, see `?ggplot2::theme`

width
The width "inch" of the plot. Default is 6.

height
The height "inch" of the plot. Default is 6.

resolution
The resolution of the plot. Parameter valid if file.type = "tiff" is used. Default is 300 (300 dpi)

col.violin
Parameter valid if violin = T. Define the color of the violin plot. Default is "gray90.

col.boxplot
Define the color for boxplot. Default is "gray70".

width.boxplot
The width of boxplots. Default is 0.2.

x.lim
The range of x-axis. Default is NULL (maximum and minimum values of the data set). New arguments can be inserted as x.lim = c(x.min, x.max).

x.breaks
The breaks to be plotted in the x-axis. Default is authomatic breaks. New arguments can be inserted as x.breaks = c(breaks)

...
Other arguments of the function

Details

Five statistics are shown in this type of plot. The lower and upper hinges correspond to the first and third quartiles (the 25th and 75th percentiles). The upper whisker extends from the hinge to the largest value no further than 1.5 * IQR from the hinge (where IQR is the inter-quartile range). The lower whisker extends from the hinge to the smallest value at most 1.5 * IQR of the hinge. Data beyond the end of the whiskers are considered outlying points.

See also

Examples

# NOT RUN {
validation <- validation.AMMIF(data_ge,
                               resp = GY,
                               gen = GEN,
                               env = ENV,
                               rep = REP,
                               nboot = 500,
                               nrepval = 2)
plot(validation)
# }