Summary a MTSI object

Summary a MTSI object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for MTSI
summary(object, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

object
The MTSI object

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(METAAB)
# Based on stability only
MTSI_MODEL = WAASB(data_ge,
                   resp = c(GY, HM),
                   gen = GEN,
                   env = ENV,
                   rep = REP)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

MTSI_index = MTSI(MTSI_MODEL)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#>     Eigenvalues Variance (%) Cum. variance (%)
#> PC1   1.6566277     82.83139          82.83139
#> PC2   0.3433723     17.16861         100.00000
#> 
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#>          FA1 Communality Uniquenesses
#> GY 0.9101175   0.8283139    0.1716861
#> HM 0.9101175   0.8283139    0.1716861
#> 
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.8283139 
#> 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G6     G8     G4     G2     G9     G5     G7    G10 
#> 0.0000 0.2303 0.4846 1.0066 1.2954 1.6022 1.8318 1.8374 1.8450 3.4178 
#> 
#> -------------------------------------------------------------------------------
#> Selection differential
#> -------------------------------------------------------------------------------
#>    Factor        Xo        Xs         SD    SDperc
#> GY    FA1 0.2504736 0.1165028 -0.1339708 -53.48700
#> HM    FA1 0.6141648 0.3732404 -0.2409244 -39.22798
#> 
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>          Xo          Xs          SD      SDperc 
#>   0.4323192   0.2448716  -0.1874476 -46.3574864 
#> 
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> G3 G1
#> -------------------------------------------------------------------------------

summary(MTSI_index)

#> -------------------- Correlation matrix used used in factor analysis -----------------
#>        GY     HM
#> GY 1.0000 0.6566
#> HM 0.6566 1.0000
#> 
#> ---------------------------- Principal component analysis -----------------------------
#>     Eigenvalues Variance.... Cum..variance....
#> PC1      1.6566        82.83             82.83
#> PC2      0.3434        17.17            100.00
#> 
#> --------------------------------- Initial loadings -----------------------------------
#>      [,1]
#> GY 0.9101
#> HM 0.9101
#> 
#> -------------------------- Loadings after varimax rotation ---------------------------
#>       FA1
#> GY 0.9101
#> HM 0.9101
#> 
#> --------------------------- Scores for genotypes-ideotype -----------------------------
#>       FA1
#> G1  1.585
#> G10 4.772
#> G2  2.957
#> G3  1.355
#> G4  2.650
#> G5  3.192
#> G6  1.839
#> G7  3.200
#> G8  2.361
#> G9  3.186
#> ID1 1.355
#> 
#> ---------------------------- Multitrait stability index ------------------------------
#>     G3     G1     G6     G8     G4     G2     G9     G5     G7    G10 
#> 0.0000 0.2303 0.4846 1.0066 1.2954 1.6022 1.8318 1.8374 1.8450 3.4178 
#> 
#> ------------------------------ Selection differential ---------------------------------
#>    Factor     Xo     Xs      SD SDperc
#> GY    FA1 0.2505 0.1165 -0.1340 -53.49
#> HM    FA1 0.6142 0.3732 -0.2409 -39.23
#> 
#> -------------------------- Mean of Selection differential -----------------------------
#>       Xo       Xs       SD   SDperc 
#>   0.4323   0.2449  -0.1874 -46.3575 
#> 
#> -------------------------------- Selected genotypes ----------------------------------
#> G3 G1
#> ------------------------------------ End of data -------------------------------------
#> 
#> 
#> 

summary(MTSI_index,
        export = TRUE,
        file.name = "my results")