Summary a WAAS.AMMI object

Summary the WAAS.AMMI object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for WAAS.AMMI
summary(object, export = FALSE, file.name = NULL,
  digits = 4, ...)

Arguments

object
The WAAS.AMMI object

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(METAAB)
model = WAAS.AMMI(data_ge,
                  resp = c(GY, HM),
                  gen = GEN,
                  env = ENV,
                  rep = REP)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

summary(model)

#> Variable GY 
#> ----------------------------- Individual analysis of variance---------------------------
#>      MEAN        MSB       MSG        MSR        FCB         PRFB       FCG
#>  2.520683 0.06522156 0.3373856 0.14400843  0.4529010 6.428307e-01  2.342818
#>  2.175209 0.65367398 0.2957869 0.02667190 24.5079604 7.275296e-06 11.089831
#>  1.368265 0.37730952 0.1509216 0.10497564  3.5942579 4.859923e-02  1.437682
#>  1.608512 0.09189509 0.3201522 0.05351799  1.7170877 2.077263e-01  5.982141
#>  2.909631 0.07673787 0.7134107 0.09937446  0.7722092 4.767029e-01  7.179014
#>  1.781933 0.10352082 0.1305017 0.07533297  1.3741768 2.783555e-01  1.732332
#>  3.179959 0.69828884 0.2066253 0.17851473  3.9116594 3.884580e-02  1.157469
#>  4.064333 0.48864413 0.3350550 0.17893617  2.7308293 9.209248e-02  1.872484
#>  3.675273 0.11645979 0.5313370 0.13769365  0.8457891 4.455822e-01  3.858835
#>  3.910500 0.21880690 0.5261193 0.06635714  3.2974130 6.023636e-02  7.928601
#>  2.663214 0.15985119 0.1346433 0.05856575  2.7294314 9.219130e-02  2.299011
#>  1.988896 0.38098564 0.3369183 0.09103066  4.1852452 3.216538e-02  3.701152
#>  2.536376 0.81677173 0.2146241 0.02781035 29.3693484 2.149427e-06  7.717420
#>  3.056600 0.58259110 0.6792532 0.11091425  5.2526263 1.596351e-02  6.124129
#>          PRFG        CV        h2        AS
#>  5.935749e-02 15.054823 0.5731637 0.7570757
#>  1.097035e-05  7.508029 0.9098273 0.9538487
#>  2.443116e-01 23.679589 0.3044360 0.5517572
#>  6.473654e-04 14.382206 0.8328358 0.9125984
#>  2.095084e-04 10.834266 0.8607051 0.9277419
#>  1.534701e-01 15.402850 0.4227435 0.6501873
#>  3.763217e-01 13.286650 0.1360461 0.3688442
#>  1.230334e-01 10.407819 0.4659499 0.6826052
#>  7.122696e-03 10.096419 0.7408544 0.8607290
#>  1.101303e-04  6.587362 0.8738743 0.9348125
#>  6.345403e-02  9.086902 0.5650303 0.7516850
#>  8.726632e-03 15.169867 0.7298139 0.8542914
#>  1.314195e-04  6.574907 0.8704230 0.9329646
#>  5.621781e-04 10.895698 0.8367115 0.9147193
#> 
#> -------------------------------------- AMMI analysis -----------------------------------
#>            Df     Sum Sq     Mean Sq   F value       Pr(>F) Percent Accumul
#> ENV        13 279.573552 21.50565785 62.325457 0.000000e+00       .       .
#> REP(ENV)   28   9.661516  0.34505416  3.568548 3.593191e-08       .       .
#> GEN         9  12.995044  1.44389374 14.932741 2.190118e-19       .       .
#> ENV:GEN   117  31.219565  0.26683389  2.759595 1.005191e-11       .       .
#> PC1        21  10.749140  0.51186000  5.290000 0.000000e+00    34.4    34.4
#> PC2        19   9.923920  0.52231000  5.400000 0.000000e+00    31.8    66.2
#> PC3        17   4.039180  0.23760000  2.460000 1.400000e-03    12.9    79.2
#> PC4        15   3.073770  0.20492000  2.120000 9.600000e-03     9.8      89
#> PC5        13   1.446440  0.11126000  1.150000 3.176000e-01     4.6    93.6
#> PC6        11   0.932240  0.08475000  0.880000 5.606000e-01       3    96.6
#> PC7         9   0.566700  0.06297000  0.650000 7.535000e-01     1.8    98.4
#> PC8         7   0.362320  0.05176000  0.540000 8.037000e-01     1.2    99.6
#> PC9         5   0.125860  0.02517000  0.260000 9.345000e-01     0.4     100
#> Residuals 252  24.366674  0.09669315        NA           NA       .       .
#> Total     419 357.816350  0.85397697        NA           NA       .       .
#> 
#> ------------------------- Weighted average of the absolute scores ----------------------
#>  type Code        Y         PC1          PC2          PC3         PC4
#>   GEN   G1 2.603676  0.31663417 -0.044167345 -0.036001128 -0.06595003
#>   GEN  G10 2.471221 -1.00066870 -0.571801414 -0.165169282 -0.33087369
#>   GEN   G2 2.744147  0.13896700  0.198761092 -0.733073122  0.47346560
#>   GEN   G3 2.955344  0.04340468 -0.102838186  0.228415475  0.17688550
#>   GEN   G4 2.641915 -0.32508656  0.478227571 -0.090731610  0.14171438
#>   GEN   G5 2.537344 -0.32597947  0.246083819  0.245219283  0.17940753
#>   GEN   G6 2.534030 -0.09836196  0.242894979  0.560735731  0.23766340
#>   GEN   G7 2.740909  0.28485465  0.587113435 -0.206799391 -0.70851977
#>   GEN   G8 3.003638  0.49947972 -0.191580376  0.319105699 -0.16764404
#>   GEN   G9 2.510192  0.46675648 -0.842693574 -0.121701653  0.06385112
#>   ENV   E1 2.520683  0.24644526  0.197617346 -0.232071269  0.09863134
#>   ENV  E10 2.175209  0.21714942  0.369574327  0.114487626  0.02614435
#>   ENV  E11 1.368265  0.08229361  0.166002067  0.152763858  0.22582137
#>   ENV  E12 1.608512  0.21475671  0.271276204 -0.181655992  0.06975203
#>   ENV  E13 2.909631  0.72492445  0.131290513  0.206894833  0.13185963
#>   ENV  E14 1.781933 -0.29124349  0.003196976  0.287008064 -0.24708939
#>   ENV   E2 3.179959 -0.22882620 -0.238363164  0.244600487  0.50779098
#>   ENV   E3 4.064333  0.06575336 -0.419722757 -0.753349309  0.01080155
#>   ENV   E4 3.675273 -0.32474063 -0.816280023 -0.002139773  0.16476991
#>   ENV   E5 3.910500  0.51053457 -0.131257930  0.093112732 -0.23980452
#>   ENV   E6 2.663214  0.05806288 -0.191721263 -0.036877134 -0.07973320
#>   ENV   E7 1.988896 -0.26583009  0.188742796 -0.187122766 -0.61812986
#>   ENV   E8 2.536376 -0.26419589 -0.226458989  0.460326020 -0.28536053
#>   ENV   E9 3.056600 -0.74508399  0.696103896 -0.165977376  0.23454632
#>          PC5         PC6         PC7           PC8          PC9      WAAS
#>  -0.31254195  0.42716876 -0.14794007  0.2543204068  0.166939384 0.1262468
#>  -0.12425037 -0.10639408 -0.03450867  0.0687603920 -0.011832060 0.5506841
#>  -0.04815789 -0.28406474 -0.07718621  0.0865630903 -0.004828096 0.3647083
#>  -0.12701680 -0.14001702  0.45680670 -0.0965877491  0.227166521 0.1305801
#>  -0.19242271  0.35495630  0.08926476 -0.2919829686 -0.174398946 0.2719624
#>   0.46624655  0.03314803 -0.29224394 -0.1211984096  0.188310907 0.2552917
#>   0.05093937 -0.10106510  0.05894710  0.3273158616 -0.201043984 0.2746953
#>   0.23146991 -0.08405567  0.12904339  0.0465900241 -0.009254332 0.4344512
#>  -0.32614955 -0.28860241 -0.27387764 -0.2222336526 -0.055782685 0.3065078
#>   0.38188341  0.18892594  0.09169459 -0.0515469950 -0.125276709 0.4005522
#>  -0.08082931 -0.17680506  0.25100893 -0.1693333409 -0.149897048 0.1993717
#>  -0.03517227  0.07490420 -0.25075175  0.2373432575  0.095571396 0.1933403
#>   0.04646607  0.05809246 -0.14181497  0.2321324416 -0.183742024 0.1509238
#>   0.35483082 -0.29468029 -0.22242384 -0.2157789070  0.031019689 0.1916468
#>   0.02074558 -0.10465712  0.33443671  0.1040240837  0.154623148 0.3224033
#>   0.19202273  0.11023698  0.15171094 -0.0786837326 -0.217193718 0.2050324
#>  -0.09064328  0.03071487  0.01470635  0.0494095574 -0.097777215 0.2931850
#>   0.26948765  0.17138159 -0.03065021  0.1216531314 -0.021907686 0.3100627
#>  -0.23756972 -0.29853597 -0.03193701 -0.0007188229  0.095361530 0.3449555
#>  -0.39991275  0.12709074 -0.26857079 -0.2331498426 -0.054108678 0.2554877
#>  -0.02537386  0.46645607  0.14340237 -0.0933340032  0.124390557 0.0928463
#>  -0.23761209 -0.21011404  0.08165507  0.2174455791 -0.019024848 0.3002997
#>   0.37817005 -0.04369699 -0.04918397 -0.0477972989  0.115412101 0.3049553
#>  -0.15460962  0.08961259  0.01841217 -0.1232121027  0.127272796 0.4890632
#>     PctResp   PctWAAS OrResp OrWAAS OrPC1 PesRes PesWAAS    WAASY OrWAASY
#>   24.878131 100.00000      6      1     5     50      50 62.43907       3
#>    0.000000   0.00000     10     10    10     50      50  0.00000      10
#>   51.261754  43.81703      3      7     3     50      50 47.53939       5
#>   90.929398  98.97903      2      2     1     50      50 94.95421       1
#>   32.060337  65.66851      5      4     6     50      50 48.86442       4
#>   12.419523  69.59623      7      3     7     50      50 41.00788       6
#>   11.797078  65.02462      8      5     2     50      50 38.41085       8
#>   50.653650  27.38516      4      9     4     50      50 39.01940       7
#>  100.000000  57.52942      1      6     9     50      50 78.76471       2
#>    7.319641  35.37198      9      8     8     50      50 21.34581       9
#>   42.744411  73.11437      9      5     7     50      50 57.92939       6
#>   29.930420  74.63661     10      4     5     50      50 52.28351       7
#>    0.000000  85.34200     14      2     3     50      50 42.67100      11
#>    8.911037  75.06403     13      3     4     50      50 41.98753      12
#>   57.170881  42.06278      6     12    13     50      50 49.61683       8
#>   15.343402  71.68568     12      6    10     50      50 43.51454      10
#>   67.197633  49.43710      4      8     6     50      50 58.31737       5
#>  100.000000  45.17740      1     11     2     50      50 72.58870       3
#>   85.569334  36.37090      3     13    11     50      50 60.97012       4
#>   94.294161  58.95143      2      7    12     50      50 76.62280       1
#>   48.031006 100.00000      7      1     1     50      50 74.01550       2
#>   23.019864  47.64145     11      9     9     50      50 35.33066      13
#>   43.326456  46.46643      8     10     8     50      50 44.89644       9
#>   62.622119   0.00000      5     14    14     50      50 31.31106      14
#> 
#> ---------------------------- Means for genotypes-vs-environments -----------------------
#>  ENV GEN        Y      envPC1      genPC1  nominal
#>   E1  G1 2.365787  0.24644526  0.31663417 2.681709
#>   E1 G10 1.974073  0.24644526 -1.00066870 2.224611
#>   E1  G2 2.901747  0.24644526  0.13896700 2.778395
#>   E1  G3 2.888587  0.24644526  0.04340468 2.966041
#>   E1  G4 2.588567  0.24644526 -0.32508656 2.561799
#>   E1  G5 2.188260  0.24644526 -0.32597947 2.457008
#>   E1  G6 2.300800  0.24644526 -0.09836196 2.509790
#>   E1  G7 2.773727  0.24644526  0.28485465 2.811110
#>   E1  G8 2.899360  0.24644526  0.49947972 3.126732
#>   E1  G9 2.325927  0.24644526  0.46675648 2.625222
#>  E10  G1 2.307633  0.21714942  0.31663417 2.672433
#>  E10 G10 1.536253  0.21714942 -1.00066870 2.253926
#>  E10  G2 2.297053  0.21714942  0.13896700 2.774324
#>  E10  G3 2.336727  0.21714942  0.04340468 2.964769
#>  E10  G4 2.167353  0.21714942 -0.32508656 2.571323
#>  E10  G5 2.139833  0.21714942 -0.32597947 2.466558
#>  E10  G6 2.208147  0.21714942 -0.09836196 2.512671
#>  E10  G7 2.441893  0.21714942  0.28485465 2.802765
#>  E10  G8 2.574867  0.21714942  0.49947972 3.112099
#>  E10  G9 1.742333  0.21714942  0.46675648 2.611548
#>  E11  G1 1.355667  0.08229361  0.31663417 2.629733
#>  E11 G10 0.899100  0.08229361 -1.00066870 2.388872
#>  E11  G2 1.490720  0.08229361  0.13896700 2.755583
#>  E11  G3 1.567727  0.08229361  0.04340468 2.958916
#>  E11  G4 1.370000  0.08229361 -0.32508656 2.615163
#>  E11  G5 1.325667  0.08229361 -0.32597947 2.510518
#>  E11  G6 1.500667  0.08229361 -0.09836196 2.525936
#>  E11  G7 1.364333  0.08229361  0.28485465 2.764351
#>  E11  G8 1.683433  0.08229361  0.49947972 3.044741
#>  E11  G9 1.125333  0.08229361  0.46675648 2.548603
#>  E12  G1 1.342333  0.21475671  0.31663417 2.671675
#>  E12 G10 1.022136  0.21475671 -1.00066870 2.256321
#>  E12  G2 1.993333  0.21475671  0.13896700 2.773991
#>  E12  G3 1.764358  0.21475671  0.04340468 2.964666
#>  E12  G4 1.527333  0.21475671 -0.32508656 2.572101
#>  E12  G5 1.689000  0.21475671 -0.32597947 2.467338
#>  E12  G6 1.385667  0.21475671 -0.09836196 2.512907
#>  E12  G7 1.951630  0.21475671  0.28485465 2.802084
#>  E12  G8 2.000000  0.21475671  0.49947972 3.110904
#>  E12  G9 1.409333  0.21475671  0.46675648 2.610431
#>  E13  G1 2.998259  0.72492445  0.31663417 2.833212
#>  E13 G10 1.830667  0.72492445 -1.00066870 1.745812
#>  E13  G2 3.028309  0.72492445  0.13896700 2.844888
#>  E13  G3 3.469148  0.72492445  0.04340468 2.986809
#>  E13  G4 2.635716  0.72492445 -0.32508656 2.406252
#>  E13  G5 2.568099  0.72492445 -0.32597947 2.301034
#>  E13  G6 2.911667  0.72492445 -0.09836196 2.462725
#>  E13  G7 3.183840  0.72492445  0.28485465 2.947407
#>  E13  G8 3.519975  0.72492445  0.49947972 3.365723
#>  E13  G9 2.950630  0.72492445  0.46675648 2.848555
#>  E14  G1 1.533333 -0.29124349  0.31663417 2.511459
#>  E14 G10 1.859210 -0.29124349 -1.00066870 2.762659
#>  E14  G2 1.426580 -0.29124349  0.13896700 2.703674
#>  E14  G3 2.059654 -0.29124349  0.04340468 2.942703
#>  E14  G4 1.861333 -0.29124349 -0.32508656 2.736595
#>  E14  G5 1.784296 -0.29124349 -0.32597947 2.632284
#>  E14  G6 1.798852 -0.29124349 -0.09836196 2.562678
#>  E14  G7 1.936333 -0.29124349  0.28485465 2.657947
#>  E14  G8 1.991864 -0.29124349  0.49947972 2.858167
#>  E14  G9 1.567877 -0.29124349  0.46675648 2.374252
#>   E2  G1 3.040690 -0.22882620  0.31663417 2.531222
#>   E2 G10 3.145840 -0.22882620 -1.00066870 2.700200
#>   E2  G2 3.231053 -0.22882620  0.13896700 2.712348
#>   E2  G3 3.608280 -0.22882620  0.04340468 2.945412
#>   E2  G4 3.190080 -0.22882620 -0.32508656 2.716304
#>   E2  G5 3.140137 -0.22882620 -0.32597947 2.611937
#>   E2  G6 3.291173 -0.22882620 -0.09836196 2.556538
#>   E2  G7 2.612677 -0.22882620  0.28485465 2.675727
#>   E2  G8 3.444797 -0.22882620  0.49947972 2.889343
#>   E2  G9 3.094863 -0.22882620  0.46675648 2.403386
#>   E3  G1 4.080333  0.06575336  0.31663417 2.624496
#>   E3 G10 4.114333  0.06575336 -1.00066870 2.405424
#>   E3  G2 4.568667  0.06575336  0.13896700 2.753285
#>   E3  G3 4.132333  0.06575336  0.04340468 2.958198
#>   E3  G4 3.854333  0.06575336 -0.32508656 2.620540
#>   E3  G5 3.741333  0.06575336 -0.32597947 2.515910
#>   E3  G6 3.434667  0.06575336 -0.09836196 2.527563
#>   E3  G7 4.101333  0.06575336  0.28485465 2.759639
#>   E3  G8 4.110000  0.06575336  0.49947972 3.036480
#>   E3  G9 4.506000  0.06575336  0.46675648 2.540883
#>   E4  G1 3.494333 -0.32474063  0.31663417 2.500852
#>   E4 G10 4.271000 -0.32474063 -1.00066870 2.796179
#>   E4  G2 3.715573 -0.32474063  0.13896700 2.699019
#>   E4  G3 4.134000 -0.32474063  0.04340468 2.941249
#>   E4  G4 3.302167 -0.32474063 -0.32508656 2.747484
#>   E4  G5 3.379113 -0.32474063 -0.32597947 2.643203
#>   E4  G6 3.403473 -0.32474063 -0.09836196 2.565973
#>   E4  G7 3.018953 -0.32474063  0.28485465 2.648405
#>   E4  G8 4.137773 -0.32474063  0.49947972 2.841436
#>   E4  G9 3.896340 -0.32474063  0.46675648 2.358617
#>   E5  G1 4.170531  0.51053457  0.31663417 2.765329
#>   E5 G10 3.365667  0.51053457 -1.00066870 1.960345
#>   E5  G2 3.827432  0.51053457  0.13896700 2.815094
#>   E5  G3 4.126654  0.51053457  0.04340468 2.977504
#>   E5  G4 3.782605  0.51053457 -0.32508656 2.475948
#>   E5  G5 3.469000  0.51053457 -0.32597947 2.370921
#>   E5  G6 3.568926  0.51053457 -0.09836196 2.483813
#>   E5  G7 4.047914  0.51053457  0.28485465 2.886337
#>   E5  G8 4.812099  0.51053457  0.49947972 3.258639
#>   E5  G9 3.934173  0.51053457  0.46675648 2.748487
#>   E6  G1 2.809086  0.05806288  0.31663417 2.622061
#>   E6 G10 2.484877  0.05806288 -1.00066870 2.413119
#>   E6  G2 2.541333  0.05806288  0.13896700 2.752216
#>   E6  G3 2.984716  0.05806288  0.04340468 2.957864
#>   E6  G4 2.701185  0.05806288 -0.32508656 2.623040
#>   E6  G5 2.433321  0.05806288 -0.32597947 2.518417
#>   E6  G6 2.335556  0.05806288 -0.09836196 2.528319
#>   E6  G7 2.665901  0.05806288  0.28485465 2.757449
#>   E6  G8 2.908123  0.05806288  0.49947972 3.032639
#>   E6  G9 2.768037  0.05806288  0.46675648 2.537293
#>   E7  G1 1.897880 -0.26583009  0.31663417 2.519505
#>   E7 G10 2.243627 -0.26583009 -1.00066870 2.737229
#>   E7  G2 1.987627 -0.26583009  0.13896700 2.707205
#>   E7  G3 2.158547 -0.26583009  0.04340468 2.943806
#>   E7  G4 1.980887 -0.26583009 -0.32508656 2.728333
#>   E7  G5 1.656767 -0.26583009 -0.32597947 2.624000
#>   E7  G6 1.757787 -0.26583009 -0.09836196 2.560178
#>   E7  G7 2.550813 -0.26583009  0.28485465 2.665186
#>   E7  G8 2.261780 -0.26583009  0.49947972 2.870861
#>   E7  G9 1.393247 -0.26583009  0.46675648 2.386114
#>   E8  G1 2.271933 -0.26419589  0.31663417 2.520023
#>   E8 G10 2.700310 -0.26419589 -1.00066870 2.735593
#>   E8  G2 2.045297 -0.26419589  0.13896700 2.707432
#>   E8  G3 2.850420 -0.26419589  0.04340468 2.943877
#>   E8  G4 2.300590 -0.26419589 -0.32508656 2.727802
#>   E8  G5 2.708330 -0.26419589 -0.32597947 2.623467
#>   E8  G6 2.539380 -0.26419589 -0.09836196 2.560017
#>   E8  G7 2.583383 -0.26419589  0.28485465 2.665652
#>   E8  G8 2.878853 -0.26419589  0.49947972 2.871677
#>   E8  G9 2.485260 -0.26419589  0.46675648 2.386877
#>   E9  G1 2.783667 -0.74508399  0.31663417 2.367757
#>   E9 G10 3.150000 -0.74508399 -1.00066870 3.216803
#>   E9  G2 3.363333 -0.74508399  0.13896700 2.640605
#>   E9  G3 3.293667 -0.74508399  0.04340468 2.923004
#>   E9  G4 3.724667 -0.74508399 -0.32508656 2.884132
#>   E9  G5 3.299667 -0.74508399 -0.32597947 2.780227
#>   E9  G6 3.039667 -0.74508399 -0.09836196 2.607318
#>   E9  G7 3.140000 -0.74508399  0.28485465 2.528669
#>   E9  G8 2.828000 -0.74508399  0.49947972 2.631483
#>   E9  G9 1.943333 -0.74508399  0.46675648 2.162419
#> 
#> ------------------------- Some information regarding the analysis ----------------------
#>  Parameters                        Values
#>        Ngen                            10
#>        Nenv                            14
#>      OVmean                        2.6742
#>         Min 0.8991 (Genotype G10 in E11 )
#>         Max   4.8121 (Genotype G8 in E5 )
#>      MinENV     Environment E11 (1.3683) 
#>      MaxENV      Environment E3 (4.0643) 
#>      MinGEN        Genotype G10 (2.4712) 
#>      MaxGEN         Genotype G8 (3.0036) 
#>       SigPC                             4
#> ------------------------------------ End of data ---------------------------------------
#> Variable HM 
#> ----------------------------- Individual analysis of variance---------------------------
#>      MEAN       MSB       MSG      MSR        FCB         PRFB       FCG
#>  47.41833  3.106343  5.220883 1.851966  1.6773224 2.147936e-01  2.819104
#>  44.30000  9.100000 28.774074 2.174074  4.1856899 3.215562e-02 13.235094
#>  54.19833  2.810083  6.057120 3.087120  0.9102604 4.201622e-01  1.962062
#>  49.59667  2.766083 10.404778 5.321917  0.5197532 6.033251e-01  1.955081
#>  46.56000 25.452250 13.883185 1.342157 18.9636848 3.705382e-05 10.343932
#>  41.03167 10.015083  5.278231 1.216843  8.2303853 2.894194e-03  4.337645
#>  44.12033  6.698603 13.168122 3.048966  2.1970080 1.400441e-01  4.318881
#>  52.90000  7.300000  8.003704 2.781481  2.6245007 9.995714e-02  2.877497
#>  50.00000  2.100000  7.925926 1.248148  1.6824926 2.138598e-01  6.350148
#>  52.23067  6.860213 11.834139 6.370861  1.0768109 3.616384e-01  1.857541
#>  45.89000 12.864000 15.835222 2.121778  6.0628404 9.705563e-03  7.463186
#>  48.49500  0.705250  4.996861 4.738028  0.1488489 8.627491e-01  1.054629
#>  45.16167 11.954083 11.567231 2.168065  5.5137112 1.355740e-02  5.335279
#>  51.33333  5.733333  9.333333 2.288889  2.5048544 1.097116e-01  4.077670
#>          PRFG       CV         h2        AS
#>  2.927958e-02 2.869922 0.64527735 0.8032916
#>  3.024284e-06 3.328384 0.92444330 0.9614797
#>  1.068927e-01 3.241835 0.49033201 0.7002371
#>  1.080680e-01 4.651377 0.48851222 0.6989365
#>  1.798467e-05 2.488220 0.90332497 0.9504341
#>  3.934627e-03 2.688426 0.76946017 0.8771888
#>  4.024648e-03 3.957651 0.76845853 0.8766177
#>  2.691721e-02 3.152698 0.65247571 0.8077597
#>  4.509981e-04 2.234411 0.84252336 0.9178907
#>  1.259610e-01 4.832518 0.46165400 0.6794513
#>  1.633196e-04 3.174182 0.86600897 0.9305960
#>  4.382354e-01 4.488506 0.05179919 0.2275943
#>  1.265758e-03 3.260365 0.81256839 0.9014258
#>  5.407546e-03 2.947222 0.75476190 0.8687703
#> 
#> -------------------------------------- AMMI analysis -----------------------------------
#>            Df     Sum Sq    Mean Sq   F value       Pr(>F) Percent Accumul
#> ENV        13 5710.31673 439.255133 57.223777 1.110223e-16       .       .
#> REP(ENV)   28  214.93065   7.676095  2.702830 2.196589e-05       .       .
#> GEN         9  269.81118  29.979019 10.555915 7.414690e-14       .       .
#> ENV:GEN   117 1100.73412   9.407984  3.312646 1.063605e-15       .       .
#> PC1        21  381.12827  18.148970  6.390000 0.000000e+00    34.6    34.6
#> PC2        19  319.43319  16.812270  5.920000 0.000000e+00      29    63.6
#> PC3        17  114.26443   6.721440  2.370000 2.100000e-03    10.4      74
#> PC4        15   81.96192   5.464130  1.920000 2.180000e-02     7.4    81.5
#> PC5        13   68.11488   5.239610  1.840000 3.770000e-02     6.2    87.7
#> PC6        11   59.07451   5.370410  1.890000 4.100000e-02     5.4      93
#> PC7         9   46.69408   5.188230  1.830000 6.330000e-02     4.2    97.3
#> PC8         7   26.65417   3.807740  1.340000 2.318000e-01     2.4    99.7
#> PC9         5    3.40867   0.681730  0.240000 9.445000e-01     0.3     100
#> Residuals 252  715.68528   2.840021        NA           NA       .       .
#> Total     419 8011.47797  19.120472        NA           NA       .       .
#> 
#> ------------------------- Weighted average of the absolute scores ----------------------
#>  type Code        Y        PC1         PC2          PC3          PC4
#>   GEN   G1 47.07976  0.2800378  0.46345799  0.173960469 -1.369244805
#>   GEN  G10 48.50643 -1.7785276  1.86581427 -0.006219455  0.921878835
#>   GEN   G2 46.65976  1.5633439  0.55184303 -0.935710557  0.491255848
#>   GEN   G3 47.60381  0.3417105 -0.20117861 -0.800098301  0.375280906
#>   GEN   G4 48.03190 -0.2020372 -1.84110603  0.280130984  0.005953625
#>   GEN   G5 49.29762  1.5795028  1.03001159  1.077783518 -0.278910540
#>   GEN   G6 48.73143  0.5473813 -0.24525280  0.532363294  0.460273871
#>   GEN   G7 47.96619 -1.2178726 -0.46798155  1.254428727 -0.054818886
#>   GEN   G8 49.10238 -0.0417645 -1.24123409 -0.410457408  0.639371003
#>   GEN   G9 47.90357 -1.0717744  0.08562619 -1.166181273 -1.191039856
#>   ENV   E1 47.41833  0.1360764 -0.16907776  0.495099364 -0.541028537
#>   ENV  E10 44.30000  2.4935426  1.01178201  0.356768228 -0.494267977
#>   ENV  E11 54.19833 -0.4653355 -0.59877405 -0.423673366 -0.036625791
#>   ENV  E12 49.59667 -0.7441848 -0.33767033  0.404641629 -0.247002930
#>   ENV  E13 46.56000 -1.1051803 -0.03001280  0.235896120  0.267835766
#>   ENV  E14 41.03167 -0.7305522  0.12973083  0.478829351 -0.439206344
#>   ENV   E2 44.12033  0.4910752  0.18732157  1.389294780  1.201257958
#>   ENV   E3 52.90000 -0.6910501  0.52931929  0.268717081  0.060741947
#>   ENV   E4 50.00000 -0.4244556  0.21439320  0.203973000  0.367306399
#>   ENV   E5 52.23067  0.3345165  1.26609303 -1.439596561  1.002509542
#>   ENV   E6 45.89000  0.9144573 -1.79069698 -0.347139653 -0.170210428
#>   ENV   E7 48.49500 -0.3285376 -0.07160880 -0.729575558  0.298759008
#>   ENV   E8 45.16167  0.5830163 -1.50122870 -0.484574158  0.017459662
#>   ENV   E9 51.33333 -0.4633883  1.16042950 -0.408660256 -1.287528275
#>          PC5          PC6          PC7          PC8          PC9      WAAS
#>  -1.13517147  0.036575962 -0.266821520  0.337572088  0.439998479 0.5556465
#>   0.10957611 -0.009744889  0.235070800  0.077996319  0.327795938 0.9194286
#>   0.28430484  1.183833486 -0.722695603  0.033680431 -0.007947627 0.8678051
#>  -0.49785970 -1.294456461 -0.498944241 -0.880324545 -0.049846822 0.5306291
#>   0.82009971  0.273370288  0.241158100 -0.437740921  0.582344376 0.6014968
#>   1.00504838 -0.736800048  0.275998135  0.162377796 -0.104567390 1.0043353
#>  -1.00827086  0.586132945  1.196183502 -0.400868469 -0.280420309 0.5381675
#>  -0.03428544  0.336625467 -1.089954547 -0.107563788 -0.414066723 0.6329493
#>  -0.17851521 -0.514885279  0.096046382  1.298946961 -0.104293595 0.5105550
#>   0.63507363  0.139348528  0.533958992 -0.084075872 -0.388996328 0.7459753
#>   0.48251760  0.086058543  0.399377502 -0.024035423 -0.023973828 0.3106414
#>  -0.22549300 -0.217574206  0.135002236 -0.005999303 -0.168530930 0.9415839
#>  -1.22739853  0.808240360  0.062125951  0.260918784 -0.574572352 0.5598689
#>  -1.26082309 -0.906877474  0.273152569 -0.578775011  0.352276938 0.6145200
#>   0.58841717 -1.088069856 -0.177418030  0.774048189 -0.234688540 0.5356770
#>   0.09734512  0.369507194 -1.422565382 -0.353155267  0.015004876 0.3944244
#>  -0.26147903  0.430603880 -0.211027223  0.517609177  0.219800932 0.6629629
#>   0.76956776  0.793984155  0.360455909 -0.802944994 -0.064551101 0.5081942
#>   0.28613739 -0.418460987  0.817326112 -0.405620272 -0.324392325 0.3154899
#>  -0.07259127 -0.363092681 -0.528567861 -0.337560617  0.017365631 0.7867615
#>   0.33968458 -0.170343711 -0.260876151 -0.269924932  0.226573638 0.7080318
#>   0.04262558  0.655896708  0.657686974  0.436631083  0.516351299 0.3428440
#>   0.40377497 -0.037154308 -0.106597290  0.185296047 -0.137265911 0.5721262
#>   0.03771475  0.057282384  0.001924686  0.603512538  0.180601672 0.6162486
#>    PctResp   PctWAAS OrResp OrWAAS OrPC1 PesRes PesWAAS    WAASY OrWAASY
#>   15.92201  90.86811      9      4     3     50      50 53.39506       6
#>   70.00632  17.19523      4      9    10     50      50 43.60077       9
#>    0.00000  27.64999     10      8     8     50      50 13.82499      10
#>   35.78843  95.93460      8      2     4     50      50 65.86152       4
#>   52.01733  81.58253      5      5     2     50      50 66.79993       3
#>  100.00000   0.00000      1     10     9     50      50 50.00000       7
#>   78.53597  94.40794      3      3     5     50      50 86.47195       2
#>   49.52613  75.21279      6      6     7     50      50 62.36946       5
#>   92.59861 100.00000      2      1     1     50      50 96.29930       1
#>   47.15227  52.32286      7      7     6     50      50 49.73756       8
#>   48.50633 100.00000      8      1     1     50      50 74.25316       5
#>   24.82278   0.00000     12     14    14     50      50 12.41139      14
#>  100.00000  60.49918      1      7     6     50      50 80.24959       2
#>   65.05063  51.83735      6      9    11     50      50 58.44399       7
#>   41.98734  64.33342      9      6    13     50      50 53.16038       9
#>    0.00000  86.72098     14      4    10     50      50 43.36049      11
#>   23.45823  44.15949     13     11     7     50      50 33.80886      13
#>   90.13924  68.68925      2      5     9     50      50 79.41425       3
#>   68.11392  99.23154      5      2     4     50      50 83.67273       1
#>   85.05570  24.53828      3     13     3     50      50 54.79699       8
#>   36.89873  37.01638     10     12    12     50      50 36.95756      12
#>   56.68354  94.89611      7      3     2     50      50 75.78983       4
#>   31.36709  58.55648     11      8     8     50      50 44.96178      10
#>   78.24051  51.56339      4     10     5     50      50 64.90195       6
#> 
#> ---------------------------- Means for genotypes-vs-environments -----------------------
#>  ENV GEN        Y     envPC1     genPC1  nominal
#>   E1  G1 46.52667  0.1360764  0.2800378 47.11787
#>   E1 G10 46.91333  0.1360764 -1.7785276 48.26441
#>   E1  G2 45.33000  0.1360764  1.5633439 46.87250
#>   E1  G3 45.88667  0.1360764  0.3417105 47.65031
#>   E1  G4 48.29333  0.1360764 -0.2020372 48.00441
#>   E1  G5 49.88333  0.1360764  1.5795028 49.51255
#>   E1  G6 48.25000  0.1360764  0.5473813 48.80591
#>   E1  G7 47.45000  0.1360764 -1.2178726 47.80047
#>   E1  G8 47.96667  0.1360764 -0.0417645 49.09670
#>   E1  G9 47.68333  0.1360764 -1.0717744 47.75773
#>  E10  G1 45.33333  2.4935426  0.2800378 47.77805
#>  E10 G10 41.66667  2.4935426 -1.7785276 44.07159
#>  E10  G2 46.33333  2.4935426  1.5633439 50.55803
#>  E10  G3 44.33333  2.4935426  0.3417105 48.45588
#>  E10  G4 41.66667  2.4935426 -0.2020372 47.52812
#>  E10  G5 51.00000  2.4935426  1.5795028 53.23618
#>  E10  G6 46.33333  2.4935426  0.5473813 50.09635
#>  E10  G7 41.00000  2.4935426 -1.2178726 44.92937
#>  E10  G8 43.66667  2.4935426 -0.0417645 48.99824
#>  E10  G9 41.66667  2.4935426 -1.0717744 45.23106
#>  E11  G1 54.00000 -0.4653355  0.2800378 46.94945
#>  E11 G10 54.00000 -0.4653355 -1.7785276 49.33404
#>  E11  G2 52.66667 -0.4653355  1.5633439 45.93228
#>  E11  G3 53.33333 -0.4653355  0.3417105 47.44480
#>  E11  G4 54.00000 -0.4653355 -0.2020372 48.12592
#>  E11  G5 51.90000 -0.4653355  1.5795028 48.56262
#>  E11  G6 56.33333 -0.4653355  0.5473813 48.47671
#>  E11  G7 54.85000 -0.4653355 -1.2178726 48.53291
#>  E11  G8 56.33333 -0.4653355 -0.0417645 49.12182
#>  E11  G9 54.56667 -0.4653355 -1.0717744 48.40231
#>  E12  G1 49.91667 -0.7441848  0.2800378 46.87136
#>  E12 G10 50.48333 -0.7441848 -1.7785276 49.82998
#>  E12  G2 44.66667 -0.7441848  1.5633439 45.49635
#>  E12  G3 50.66667 -0.7441848  0.3417105 47.34951
#>  E12  G4 49.66667 -0.7441848 -0.2020372 48.18226
#>  E12  G5 49.13333 -0.7441848  1.5795028 48.12218
#>  E12  G6 51.21667 -0.7441848  0.5473813 48.32408
#>  E12  G7 50.41667 -0.7441848 -1.2178726 48.87251
#>  E12  G8 50.66667 -0.7441848 -0.0417645 49.13346
#>  E12  G9 49.13333 -0.7441848 -1.0717744 48.70117
#>  E13  G1 44.40000 -1.1051803  0.2800378 46.77027
#>  E13 G10 49.15000 -1.1051803 -1.7785276 50.47202
#>  E13  G2 42.33333 -1.1051803  1.5633439 44.93198
#>  E13  G3 46.15000 -1.1051803  0.3417105 47.22616
#>  E13  G4 46.51667 -1.1051803 -0.2020372 48.25519
#>  E13  G5 47.66667 -1.1051803  1.5795028 47.55198
#>  E13  G6 45.16667 -1.1051803  0.5473813 48.12647
#>  E13  G7 47.90000 -1.1051803 -1.2178726 49.31216
#>  E13  G8 49.20000 -1.1051803 -0.0417645 49.14854
#>  E13  G9 47.11667 -1.1051803 -1.0717744 49.08808
#>  E14  G1 40.73333 -0.7305522  0.2800378 46.87518
#>  E14 G10 42.23333 -0.7305522 -1.7785276 49.80574
#>  E14  G2 39.35000 -0.7305522  1.5633439 45.51766
#>  E14  G3 40.21667 -0.7305522  0.3417105 47.35417
#>  E14  G4 41.01667 -0.7305522 -0.2020372 48.17950
#>  E14  G5 41.23333 -0.7305522  1.5795028 48.14371
#>  E14  G6 39.85000 -0.7305522  0.5473813 48.33154
#>  E14  G7 44.06667 -0.7305522 -1.2178726 48.85591
#>  E14  G8 40.63333 -0.7305522 -0.0417645 49.13289
#>  E14  G9 40.98333 -0.7305522 -1.0717744 48.68656
#>   E2  G1 42.57333  0.4910752  0.2800378 47.21728
#>   E2 G10 45.14333  0.4910752 -1.7785276 47.63304
#>   E2  G2 43.45667  0.4910752  1.5633439 47.42748
#>   E2  G3 42.31667  0.4910752  0.3417105 47.77162
#>   E2  G4 43.77000  0.4910752 -0.2020372 47.93269
#>   E2  G5 46.88333  0.4910752  1.5795028 50.07327
#>   E2  G6 46.27333  0.4910752  0.5473813 49.00023
#>   E2  G7 45.22667  0.4910752 -1.2178726 47.36812
#>   E2  G8 45.53333  0.4910752 -0.0417645 49.08187
#>   E2  G9 40.02667  0.4910752 -1.0717744 47.37725
#>   E3  G1 50.66667 -0.6910501  0.2800378 46.88624
#>   E3 G10 55.66667 -0.6910501 -1.7785276 49.73548
#>   E3  G2 51.33333 -0.6910501  1.5633439 45.57941
#>   E3  G3 51.00000 -0.6910501  0.3417105 47.36767
#>   E3  G4 53.33333 -0.6910501 -0.2020372 48.17152
#>   E3  G5 54.00000 -0.6910501  1.5795028 48.20610
#>   E3  G6 53.66667 -0.6910501  0.5473813 48.35316
#>   E3  G7 53.66667 -0.6910501 -1.2178726 48.80780
#>   E3  G8 51.66667 -0.6910501 -0.0417645 49.13124
#>   E3  G9 54.00000 -0.6910501 -1.0717744 48.64422
#>   E4  G1 47.66667 -0.4244556  0.2800378 46.96090
#>   E4 G10 52.00000 -0.4244556 -1.7785276 49.26133
#>   E4  G2 47.00000 -0.4244556  1.5633439 45.99619
#>   E4  G3 49.66667 -0.4244556  0.3417105 47.45877
#>   E4  G4 50.00000 -0.4244556 -0.2020372 48.11766
#>   E4  G5 51.66667 -0.4244556  1.5795028 48.62719
#>   E4  G6 51.33333 -0.4244556  0.5473813 48.49909
#>   E4  G7 49.66667 -0.4244556 -1.2178726 48.48312
#>   E4  G8 50.66667 -0.4244556 -0.0417645 49.12011
#>   E4  G9 50.33333 -0.4244556 -1.0717744 48.35849
#>   E5  G1 50.38333  0.3345165  0.2800378 47.17344
#>   E5 G10 55.20000  0.3345165 -1.7785276 47.91148
#>   E5  G2 53.78333  0.3345165  1.5633439 47.18273
#>   E5  G3 54.20000  0.3345165  0.3417105 47.71812
#>   E5  G4 49.25000  0.3345165 -0.2020372 47.96432
#>   E5  G5 53.43333  0.3345165  1.5795028 49.82599
#>   E5  G6 51.80000  0.3345165  0.5473813 48.91454
#>   E5  G7 49.73333  0.3345165 -1.2178726 47.55879
#>   E5  G8 52.60000  0.3345165 -0.0417645 49.08841
#>   E5  G9 51.92333  0.3345165 -1.0717744 47.54505
#>   E6  G1 44.16667  0.9144573  0.2800378 47.33584
#>   E6 G10 41.21667  0.9144573 -1.7785276 46.88004
#>   E6  G2 45.21667  0.9144573  1.5633439 48.08937
#>   E6  G3 46.70000  0.9144573  0.3417105 47.91629
#>   E6  G4 49.26667  0.9144573 -0.2020372 47.84715
#>   E6  G5 46.70000  0.9144573  1.5795028 50.74201
#>   E6  G6 46.50000  0.9144573  0.5473813 49.23199
#>   E6  G7 45.21667  0.9144573 -1.2178726 46.85250
#>   E6  G8 48.75000  0.9144573 -0.0417645 49.06419
#>   E6  G9 45.16667  0.9144573 -1.0717744 46.92348
#>   E7  G1 47.00000 -0.3285376  0.2800378 46.98776
#>   E7 G10 50.00000 -0.3285376 -1.7785276 49.09074
#>   E7  G2 47.66667 -0.3285376  1.5633439 46.14614
#>   E7  G3 47.00000 -0.3285376  0.3417105 47.49154
#>   E7  G4 48.91667 -0.3285376 -0.2020372 48.09828
#>   E7  G5 48.00000 -0.3285376  1.5795028 48.77869
#>   E7  G6 49.53333 -0.3285376  0.5473813 48.55159
#>   E7  G7 47.11667 -0.3285376 -1.2178726 48.36631
#>   E7  G8 50.33333 -0.3285376 -0.0417645 49.11610
#>   E7  G9 49.38333 -0.3285376 -1.0717744 48.25569
#>   E8  G1 43.08333  0.5830163  0.2800378 47.24303
#>   E8 G10 41.75000  0.5830163 -1.7785276 47.46952
#>   E8  G2 44.43333  0.5830163  1.5633439 47.57122
#>   E8  G3 45.31667  0.5830163  0.3417105 47.80303
#>   E8  G4 47.75000  0.5830163 -0.2020372 47.91411
#>   E8  G5 45.66667  0.5830163  1.5795028 50.21849
#>   E8  G6 45.65000  0.5830163  0.5473813 49.05056
#>   E8  G7 44.55000  0.5830163 -1.2178726 47.25615
#>   E8  G8 48.41667  0.5830163 -0.0417645 49.07803
#>   E8  G9 45.00000  0.5830163 -1.0717744 47.27871
#>   E9  G1 52.66667 -0.4633883  0.2800378 46.95000
#>   E9 G10 53.66667 -0.4633883 -1.7785276 49.33058
#>   E9  G2 49.66667 -0.4633883  1.5633439 45.93533
#>   E9  G3 49.66667 -0.4633883  0.3417105 47.44546
#>   E9  G4 49.00000 -0.4633883 -0.2020372 48.12553
#>   E9  G5 53.00000 -0.4633883  1.5795028 48.56570
#>   E9  G6 50.33333 -0.4633883  0.5473813 48.47778
#>   E9  G7 50.66667 -0.4633883 -1.2178726 48.53054
#>   E9  G8 51.00000 -0.4633883 -0.0417645 49.12173
#>   E9  G9 53.66667 -0.4633883 -1.0717744 48.40022
#> 
#> ------------------------- Some information regarding the analysis ----------------------
#>  Parameters                        Values
#>        Ngen                            10
#>        Nenv                            14
#>      OVmean                       48.0883
#>         Min   39.35 (Genotype G2 in E14 )
#>         Max 56.3333 (Genotype G6 in E11 )
#>      MinENV    Environment E14 (41.0317) 
#>      MaxENV    Environment E11 (54.1983) 
#>      MinGEN        Genotype G2 (46.6598) 
#>      MaxGEN        Genotype G5 (49.2976) 
#>       SigPC                             6
#> ------------------------------------ End of data ---------------------------------------