Make a two-way table based on categorical and numerical arguments

This function help users to easely make a two-way table from a "long format" data.

mmat(data, row, col, value)

Arguments

data
The dataset. Must contains at least two categorical columns.

row
The column of data in which the mean of each level will correspond to one line in the output.

col
The column of data in which the mean of each level will correspond to one column in the output.

value
The column of data that contains the values to fill the two-way table.

Examples

datage = expand.grid(ENV = paste("E", 1:8, sep = ""),
                     GEN = paste("G", 1:20, sep = ""),
                     REP = paste(c("I", "II", "III")))
datage$Y = rnorm(480, 50, 20)
matrix = mmat(datage, row = GEN, col = ENV, val = Y)
matrix

#>           E1       E2       E3       E4       E5       E6       E7       E8
#> G1  32.33619 60.23343 41.54145 52.19681 56.38270 42.55971 33.02033 60.06870
#> G2  48.86794 55.93009 52.63698 34.50238 80.29389 44.40397 60.50785 52.92840
#> G3  35.79171 33.60673 52.31579 64.33733 44.61432 58.68744 29.39473 55.80579
#> G4  58.66933 74.00183 42.19972 64.07451 50.77330 52.77839 59.86838 47.56040
#> G5  49.64588 68.52162 72.16323 60.11459 37.21603 67.91064 52.50884 55.42430
#> G6  45.31216 61.89476 54.61491 58.53231 35.95366 75.35572 31.56480 44.08341
#> G7  44.36688 53.75594 49.25000 41.45131 47.46551 35.15494 51.55417 61.84854
#> G8  41.26334 67.51361 37.02152 65.87496 57.12236 42.74833 57.03737 49.42114
#> G9  38.06754 51.85580 59.42046 53.93208 53.41755 30.72661 39.03034 49.37065
#> G10 76.16444 36.46239 59.31905 62.79043 39.46577 61.49525 57.48232 35.63426
#> G11 68.38426 61.80519 55.91606 68.51655 26.82945 58.72418 36.80684 62.45115
#> G12 51.74387 37.41556 47.22355 49.35335 44.57025 41.54390 71.74298 59.87228
#> G13 52.99827 41.43280 60.70396 59.24339 52.28855 34.02283 47.82618 36.63186
#> G14 12.65605 36.17550 54.24542 27.34057 51.33882 63.95815 52.03896 20.48198
#> G15 54.71253 55.48472 48.80192 73.43557 59.45042 40.30509 57.54489 43.31052
#> G16 60.95821 51.59720 54.76234 52.53815 56.24226 55.52865 39.16599 42.83431
#> G17 66.00825 51.03547 53.06104 56.74047 37.68997 33.94181 53.83624 29.02831
#> G18 34.88228 51.56074 67.76969 63.05596 53.32919 35.63004 80.74776 57.82619
#> G19 51.69607 54.30562 42.68696 65.77647 60.36898 53.90106 45.06424 37.63271
#> G20 40.40402 19.68933 50.46697 51.51124 54.24341 44.79646 51.41667 47.79862