Summary a ge_stats object

Summary the ge_stats object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ge_stats
summary(object, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

object
The ge_stats object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(METAAB)
model = ge_stats(.data = data_ge2,
                 env = ENV,
                 gen = GEN,
                 rep = REP,
                 resp = PH)
summary(model)

#> Variable PH 
#> ------------------------------ Individual analysis of variance--- ---------------------
#> data frame with 0 columns and 0 rows
#> 
#> ------------------------------- Genotype-vs-environment mean --------------------------
#>        A1    A2    A3    A4
#> H1  2.723 2.930 2.197 2.635
#> H10 2.783 2.049 2.038 2.389
#> H11 2.749 2.147 2.101 2.563
#> H12 2.693 2.093 2.430 2.534
#> H13 2.772 2.227 2.599 2.557
#> H2  2.792 2.947 2.154 2.524
#> H3  2.935 2.941 2.043 2.460
#> H4  2.869 2.846 2.049 2.556
#> H5  2.835 2.705 2.095 2.638
#> H6  2.769 2.825 2.153 2.485
#> H7  2.774 2.137 2.177 2.529
#> H8  2.692 1.959 2.100 2.555
#> H9  2.925 2.199 2.028 2.305
#> 
#> ------------------------------ Genotype-vs-environment effects -------------------------
#>           A1      A2       A3       A4
#> H1  -0.20692  0.3316 -0.10575 -0.01891
#> H10  0.16025 -0.2426  0.04142  0.04092
#> H11  0.05041 -0.2204  0.02958  0.14042
#> H12 -0.05309 -0.3213  0.31075  0.06359
#> H13 -0.07492 -0.2884  0.37825 -0.01491
#> H2  -0.12042  0.3654 -0.13192 -0.11308
#> H3   0.03221  0.3687 -0.23328 -0.16764
#> H4  -0.01959  0.2889 -0.21242 -0.05691
#> H5  -0.04159  0.1596 -0.15508  0.03709
#> H6  -0.09759  0.2896 -0.08642 -0.10558
#> H7   0.06131 -0.2447  0.09128  0.09212
#> H8   0.05718 -0.3443  0.09168  0.19546
#> H9   0.25275 -0.1421 -0.01808 -0.09258
#> 
#> ------------------------ Genotype + Genotype-vs-environment effects --------------------
#>            A1      A2       A3        A4
#> H1  -0.070400  0.4681  0.03077  0.117610
#> H10 -0.009733 -0.4126 -0.12856 -0.129056
#> H11 -0.044400 -0.3152 -0.06523  0.045610
#> H12 -0.100400 -0.3686  0.26344  0.016277
#> H13 -0.021067 -0.2346  0.43210  0.038944
#> H2  -0.001067  0.4848 -0.01256  0.006277
#> H3   0.142267  0.4788 -0.12323 -0.057590
#> H4   0.075600  0.3841 -0.11723  0.038277
#> H5   0.041600  0.2428 -0.07190  0.120277
#> H6  -0.024400  0.3628 -0.01323 -0.032390
#> H7  -0.019200 -0.3252  0.01077  0.011610
#> H8  -0.101067 -0.5026 -0.06656  0.037210
#> H9   0.132267 -0.2626 -0.13856 -0.213056
#> 
#> ----------------------------- Genotype-vs-environment statistics -----------------------
#>      Mean Variance  GENSS Ecoval Ecov.perc
#> H1  2.621   0.2855 0.8565 0.4929     9.133
#> H10 2.315   0.3721 1.1162 0.2637     4.887
#> H11 2.390   0.3012 0.9037 0.2152     3.986
#> H12 2.438   0.1929 0.5788 0.6199    11.485
#> H13 2.539   0.1553 0.4659 0.6963    12.901
#> H2  2.604   0.3617 1.0850 0.5347     9.906
#> H3  2.595   0.5578 1.6735 0.6585    12.201
#> H4  2.580   0.4363 1.3088 0.3966     7.349
#> H5  2.568   0.3187 0.9562 0.1579     2.925
#> H6  2.558   0.2845 0.8536 0.3360     6.225
#> H7  2.404   0.2754 0.8261 0.2414     4.473
#> H8  2.327   0.3719 1.1157 0.5053     9.362
#> H9  2.364   0.4586 1.3759 0.2789     5.167
#> 
#> ----------------------------- Regression-based stability (anova) -----------------------
#>                      Df  Sum Sq Mean Sq F value Pr(>F)
#> Total                51 14.9805 0.29374      NA     NA
#> GEN                  12  1.8647 0.15539  0.8703 0.5847
#> ENV + (GEN x ENV)    39 13.1158 0.33630      NA     NA
#> ENV (linear)          1  7.7185 7.71852      NA     NA
#>  GEN x ENV (linear)  12  0.7550 0.06292  0.3524 0.9691
#> Pooled deviation     26  4.6423 0.17855      NA     NA
#> H1                    2  0.4706 0.23531  0.7240 0.4873
#> H10                   2  0.2356 0.11778  0.3624 0.6969
#> H11                   2  0.2114 0.10569  0.3252 0.7231
#> H12                   2  0.4502 0.22510  0.6926 0.5026
#> H13                   2  0.4103 0.20515  0.6312 0.5340
#> H2                    2  0.5339 0.26693  0.8213 0.4427
#> H3                    2  0.5838 0.29192  0.8981 0.4105
#> H4                    2  0.3540 0.17698  0.5445 0.5818
#> H5                    2  0.1402 0.07012  0.2157 0.8063
#> H6                    2  0.3335 0.16677  0.5131 0.6001
#> H7                    2  0.2414 0.12068  0.3713 0.6908
#> H8                    2  0.5052 0.25259  0.7771 0.4624
#> H9                    2  0.1723 0.08613  0.2650 0.7677
#> Pooled error        104 33.8026 0.32503      NA     NA
#> 
#> --------------------------- Regression-based stability (parameters) ---------------------
#>    GEN    bij     sdij
#> 1   H1 0.8062 0.062787
#> 2  H10 1.2179 0.023611
#> 3  H11 1.0798 0.019580
#> 4  H12 0.4654 0.059383
#> 5  H13 0.3060 0.052734
#> 6   H2 0.9635 0.073328
#> 7   H3 1.3547 0.081656
#> 8   H4 1.2682 0.043343
#> 9   H5 1.1723 0.007724
#> 10  H6 0.9359 0.039942
#> 11  H7 0.9924 0.024577
#> 12  H8 1.0140 0.068548
#> 13  H9 1.4238 0.013061
#> 
#> -------------------------------------- End of data -------------------------------------
#> 
#> 
#>